moved get experiment attributes back into helper
