+++ src/api/general-services/pipeline-manage/index.js
@@ -155,7 +155,10 @@ const createQCPipeline = async (experimentId, processingConfigUpdates) => {
   const accountId = await config.awsAccountIdPromise;
   const roleArn = `arn:aws:iam::${accountId}:role/state-machine-role-${config.clusterEnv}`;
   logger.log(`Fetching processing settings for ${experimentId}`);
-  const { processingConfig, sampleIds } = await experimentService.getExperimentAttributes(experimentId, ['processingConfig', 'sampleIds']);
+  const {
+    processingConfig,
+    sampleIds,
+  } = await experimentService.getAttributesToCreateQCPipeline(experimentId);
 
   if (processingConfigUpdates) {
     processingConfigUpdates.forEach(({ name, body }) => {
