+++ src/api/route-services/experiment.js
@@ -11,6 +11,7 @@ const safeBatchGetItem = require('../../utils/safeBatchGetItem');
 const constants = require('../general-services/pipeline-manage/constants');
 
 const {
+  getExperimentAttributes,
   getDeepAttrsUpdateParams,
   getShallowAttrsUpdateParams,
 } = require('./experimentHelpers');
@@ -31,31 +32,8 @@ class ExperimentService {
     this.mockData = convertToDynamoDbRecord(mockData);
   }
 
-
-  async getExperimentAttributes(experimentId, attributes) {
-    const dynamodb = createDynamoDbInstance();
-    const key = convertToDynamoDbRecord({ experimentId });
-
-    const params = {
-      TableName: this.experimentsTableName,
-      Key: key,
-    };
-
-    if (Array.isArray(attributes) && attributes.length > 0) {
-      params.ProjectionExpression = attributes.join();
-    }
-
-    const data = await dynamodb.getItem(params).promise();
-    if (Object.keys(data).length === 0) {
-      throw new NotFoundError('Experiment does not exist.');
-    }
-
-    const prettyData = convertToJsObject(data.Item);
-    return prettyData;
-  }
-
   async getExperimentData(experimentId) {
-    const data = await this.getExperimentAttributes(experimentId,
+    const data = await getExperimentAttributes(this.experimentsTableName, experimentId,
       ['projectId', 'meta', 'experimentId', 'experimentName', 'sampleIds']);
     return data;
   }
@@ -189,17 +167,17 @@ class ExperimentService {
   }
 
   async getExperimentPermissions(experimentId) {
-    const data = await this.getExperimentAttributes(experimentId, ['experimentId', 'rbac_can_write']);
+    const data = await getExperimentAttributes(this.experimentsTableName, experimentId, ['experimentId', 'rbac_can_write']);
     return data;
   }
 
   async getProcessingConfig(experimentId) {
-    const data = await this.getExperimentAttributes(experimentId, ['processingConfig']);
+    const data = await getExperimentAttributes(this.experimentsTableName, experimentId, ['processingConfig']);
     return data;
   }
 
   async getPipelinesHandles(experimentId) {
-    const data = await this.getExperimentAttributes(experimentId, ['meta']);
+    const data = await getExperimentAttributes(this.experimentsTableName, experimentId, ['meta']);
 
     return {
       [constants.QC_PROCESS_NAME]: {
@@ -213,6 +191,11 @@ class ExperimentService {
         ...data.meta[constants.GEM2S_PROCESS_NAME],
       },
     };
+  }
+
+  async getAttributesToCreateQCPipeline(experimentId) {
+    const data = await getExperimentAttributes(this.experimentsTableName, experimentId, ['processingConfig', 'sampleIds']);
+    return data;
   }
 
   async getCellSets(experimentId) {
@@ -233,7 +216,7 @@ class ExperimentService {
       if (e.code === 'NoSuchKey') {
         logger.log(`ERROR: Couldn't find s3 cell sets bucket with key: ${experimentId}`);
 
-        const actualData = await this.getExperimentAttributes(experimentId, ['cellSets']);
+        const actualData = await getExperimentAttributes(this.experimentsTableName, experimentId, ['cellSets']);
 
         if (actualData) {
           logger.log('Found the cell sets in dynamodb, this means this experiment has an OUTDATED structure and its cell sets should be moved to s3');
