+++ src/api/route-services/experimentHelpers.js
@@ -1,7 +1,33 @@
 const {
+  createDynamoDbInstance,
+  convertToJsObject,
   convertToDynamoDbRecord,
   batchConvertToDynamoUpdateParams,
 } = require('../../utils/dynamoDb');
+
+const { NotFoundError } = require('../../utils/responses');
+
+const getExperimentAttributes = async (tableName, experimentId, attributes) => {
+  const dynamodb = createDynamoDbInstance();
+  const key = convertToDynamoDbRecord({ experimentId });
+
+  const params = {
+    TableName: tableName,
+    Key: key,
+  };
+
+  if (Array.isArray(attributes) && attributes.length > 0) {
+    params.ProjectionExpression = attributes.join();
+  }
+
+  const data = await dynamodb.getItem(params).promise();
+  if (Object.keys(data).length === 0) {
+    throw new NotFoundError('Experiment does not exist.');
+  }
+
+  const prettyData = convertToJsObject(data.Item);
+  return prettyData;
+};
 
 const toUpdatePropertyArray = (updatePropertyObject) => (
   Object.entries(updatePropertyObject).map(([key, val]) => ({ name: key, body: val }))
@@ -67,6 +93,7 @@ const getShallowAttrsUpdateParams = (body) => {
 };
 
 module.exports = {
+  getExperimentAttributes,
   toUpdatePropertyArray,
   getDeepAttrsUpdateParams,
   getShallowAttrsUpdateParams,
